pyproject.toml
[project]
name = "bacscancode"
version = "1.0.0"
description = "Browser-based crawler and vulnerability scanner with Playwright."
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "treelib>=1.7.0",
    "elasticsearch>=8.0.0,<9.0.0",
    "python-Levenshtein>=0.25.0",
    "lxml>=5.0.0",
]

[tool.ruff]
target-version = 'py311'

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 120
indent-width = 4

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "D",  # pydocstyle
    "E",  # Error
    "F",  # pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "PGH",  # pygrep-hooks
    "PTH", # flake8-use-pathlib
    "Q",  # flake8-quotes
    "S",  # bandit
    "SIM",  # flake8-simplify
    "TRY",  # tryceratops
    "UP",  # pyupgrade
    "W",  # Warning
    "YTT", # flake8-2020
]

ignore = [
    "B006",
    "B905",  # zip strict=True; remove once python <3.10 support is dropped.
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D200",
    "D205",
    "D400",
    "D401",
    "E402",
    "E501",
    "E711",
    "E722",
    "F401",
    "F403",
    "F405",
    "N801",
    "N806",
    "S110",
    "S501",
    "W291",
    "W293",
    "PTH103",
    "PTH110",
    "PTH118",
    "PTH120",
    "PTH123",
    "SIM102",
    "TRY003",  # Avoid specifying messages outside exception class; overly strict, especially for ValueError
    "TRY400",
    "TRY401",
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.pep8-naming]
staticmethod-decorators = [
    "pydantic.validator",
    "pydantic.root_validator",
]

[tool.ruff.lint.per-file-ignores]
"test*" = [
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "S101",  # use of "assert"
    "S102",  # use of "exec"
    "S106",  # possible hardcoded password.
]
"__init__.py" = ["F401"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["crawler", "vuln_detection", "config"]

[dependency-groups]
dev = [
    "ruff>=0.15.4",
]
